requirements.txt
flask
flask_sqlalchemy
flask_login
psycopg2-binary
gunicorn
werkzeug
python-dateutil
pandas
openpyxl